+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add sponge over fixed input length for merkle tree hashing [#215]
+
 ## [0.29.0] - 2023-05-17
 
 ### Removed
@@ -373,6 +377,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Variants of sponge for `Scalar` & `Gadget(Variable/LC)`.
 
 <!-- ISSUES -->
+[#215]: https://github.com/dusk-network/poseidon252/issues/215
 [#212]: https://github.com/dusk-network/poseidon252/issues/212
 [#198]: https://github.com/dusk-network/poseidon252/issues/198
 [#206]: https://github.com/dusk-network/poseidon252/issues/206